Add TableService to host and join tables
This commit adds an initial implementation of the TableService and views that make use of it via the Multipeer Connectivity framework.

This service is still not in an acceptable prototype state. The largest bug is that sessions are dropped when navigating between views. The path forward seems to be to embed data-based view inside a single service controller view that will display the correct UI view based on its state.